Thermal Control Update
The Catia design of the Louvres has been added (36% of the surface of the S/C must be covered in these, taking 3m^2 as the area of each of this structures).

Additionally, some of the simulations used to make the graphs for themal control have been added to Simulations folder.